Fix symlink creation
Previously, symlinks pointed to /var/lib/amavis/tmp/amavis-... which is wrong,
since the files may not be there anymore when Peekaboo has finished its
analyses. Therefore, AMaViSd also puts the fiels in /tmp/amavis-... and we
create symlinks in this directory that point to the files in this directory.