Store information about malicious emails
If an email or file is malicious, processing information should be saved.

The following files should be saved in a directory with the name of the
current job hash:
* Cuckoo report (HTML + JSON)
* meta info file

The name for those files is alaways the SHA-256 hash of the file
that has been analyzed
e.g. 89cd0ff4eea714cef6b5500136bff9b789c023a20cd42121152692129772932f.html
for the HTML report of Cuckoo.

The base direcotry should be in the home of the user running Peekaboo
named malware_reports.