Implement new database schema
With this commit Peekaboo uses a new database schema with the version code 2.
This schema allows saving every analysis job to the database and references
the corresponding sample to it. Also, analysis results are kept in a
separate table and referenced by samples.
Note that one analysis can have multiple jobs.