Fix custom protocol call parameter grafting
Changing @_ is a bad idea because Perl is call-by-reference and changing
an element of @_ doesn't replace the reference in that array element
with another but changes the referenced object's value. In our case, due
to the internals of amavis, this would make the next call that should go
to ask_peekaboo go completely around it and run_av and directly into
ask_peekaboo_internal which would choke on the wrong argument data
types. The seeming nondeterminism of this is aggravated by the fact that
amavis' children and objects seem to have a lifetime after which they
would revert to a sane state, masking the problem.