Enhance shutdown logic
Enhance the shutdown mechanisms in various ways: First, make the signal
handler do nothing but call back into some registered objects interested
in shutdown requests. They in turn also do nothing but remember that a
shutdown action is requested in keeping with the recommendation that a
signal handler should not do lengthy work.

Move the job queue and worker shutdown logic into the job queue. With
this change the job queue is now totally governing the worker life cycle
from creation to scheduling to shutdown.

The cuckoo classes are extended as well to notice shutdown requests.
This is particularly beneficial for embedded mode because before it
would simply make the process exit on any error without giving it any
chance for cleanup. Also, looking into this releaved that twisted is
by default also installing signal handlers for SIGINT, SIGTERM and
SIGCHLD potentially interfering in unforeseen ways and causing confusion
when debugging problems.