Kill global Singletons
Convert the JobQueue and ConnectionMap global Singletons to standard
classes of which we create one instance each and hand them as references
to all other objects employing them. The main reason is to highlight
architectural breakage and impose serious pain on any programmer intent
on creating any. As a minor side-effect this avoids importing the
modules everywhere because user objects can just use them without
knowing anything about them other than what method to call.

As specific improvement, move the logic to create the workers into the
JobQueue now that we govern the time when it's instantiated.