Kill the global config object
Having a global configuration object accessible from everywhere obscures
who is using which configuration data and what is actually configurable
and what isn't. Addtionally, in the case of the database connection and
cuckoo wrapper object it was abused as basically a global variable,
allowing arbitrary objects to be accessible everywhere, obscuring
architectural boundaries.

Remove the global config object. Have just one instance in the main
daemon. Do not hand it to anyone as a whole. Instead hand individual
configuration items as parameters to objects via their constructor,
overriding their internal defaults, thus providing a very clear and
visible documentation of what is adjustable from the outside and what
isn't.

As with the global Singletons it is the declared intention of this
change to highlight architectural breakage by imposing pain on the
programmer wishing to access some config item in the very last corner of
the code base.

One special case is currently the Sample class which needs so many
parameters that we resort to basically a mini Sample-Configuration class
called SampleFactory holding all the items needed by but identical for
all samples. The goal here is to eventually simplify this or refactor at
least some of those out of Sample.