Simplify and fix cuckoo report handling
Move acquiring the cuckoo report from the CuckooReport into the Cuckoo
classes. This eliminates a need for explicitly distinguishing the mode
again. Incidentally this fixes that the mode wasn't switched in api mode
and CuckooReport would still access the files in the storage directory
instead of downloading the report via the API.

Do not copy the files but reserialise the loaded report as json. Drop
copying the HTML report for now. If we need it, it can be downloaded via
the REST API as well. The problem here is whether we would load it into
memory or create a temporary directory for it. Both options aren't all
that appealing.