Remove cuckoo report wrapper functions
Remove the wrappers for Cuckoo report's failed analysis and requested
domains indicators and use the cuckoo_report property of Sample
directly. This restores the self-organising property of samples in that
the first call to cuckoo_report will trigger the submit to Cuckoo,
allowing for reordering of Cuckoo rules accessing the report without
changes to Sample. Also, it removes some boilerplate code as well as
otherwise unused sample attributes.