Avoid startup slowdown due to orphaned DB entries
When starting up we remove entries from the analysis journal which
reference samples which have not finished analysis on the previous run.
We do however not remove those samples themselves from the sample info
table. That causes them to accumulate there. This in turn will dradually
slow down startup because on each startup we not only delete the
unfinished analysis journal entries left from the previous run but also
those which have already been deleted earlier.

Also, if we were using the sample info in progress state of the db to
decide whether to wait for it to finish (which we're apparently not
doing right now), leaving those entries would cause endless waiting for
analysis which will never finish because they've been interrupted by
shutdown of a previous run.