Handle being unable to get the Cuckoo report
When the Cuckoo report couldn't be retrieved, e.g because the file was
missing or the JSON was invalid, the cuckoo_report Sample attribute
would still be set but to None. This would cause errors further down the
ruleset such as:

peekaboo.ruleset.engine - WARNING - Unexpected error in 'cuckoo_evil_sig' for <Sample(...)>
peekaboo.ruleset.engine - ERROR - 'NoneType' object has no attribute '__getitem__'
Traceback (most recent call last):
  File "build/bdist.linux-x86_64/egg/peekaboo/ruleset/engine.py", line 113, in __exec_rule
    result = rule_function(config, sample)
  File "build/bdist.linux-x86_64/egg/peekaboo/ruleset/rules.py", line 128, in cuckoo_evil_sig
    for descr in s.cuckoo_report.signatures:
  File "build/bdist.linux-x86_64/egg/peekaboo/toolbox/cuckoo.py", line 229, in signatures
    return self.report['signatures']
TypeError: 'NoneType' object has no attribute '__getitem__'

With this change we do not set the cuckoo_report Sample attribute so
that it will be resubmitted to Cuckoo upon the next access to the
report. This might cause an infinite loop though if we persistently fail
to access the Cuckoo report. A retry counter would be needed here to
avoid that.