Handle case of not knowing any mimetype
Previously we would treat a sample as whitelisted if we would not know any mimetype for it.
This potentially leaves samples unanalysed.
Therefore we check if the set of known mimetypes for the sample is non empty.

We also needed this change for the graylist to analyse files with no determined mimetype.