Revert "Avoid startup slowdown due to orphaned DB entries"
When running multiple instances of Peekaboo with the same database,
these changes cause records of all currently running analysis to be
removed on restart/startup of a single instance. This causes all
currently running analysis of the other instances to fail. We need some
improved code to track this distinction. Until then we should back out
this change.

This reverts commit 8c1440d7465163248bd9f40ee84dc4f7a5900deb.