Implement simple instance concurrency
The aim is to have multiple Peekaboo instances to run in parallel to
achieve higher throughput. Duplicate concurrent analysis of the same
sample is to be avoided. To that end we:

Give each Peekaboo instance a unique id. Replace special analysis
"result" inProgress with a separate table where we track each sample by
checksum. There we also track the instance processing it. Record
in-flight samples using a (hopefully) atomic INSERT operation that either
succeeds or fails, thus reliably indicating whether we now "own" this
sample for processing or another instance is currently working on it.

Extend the in-flight logic in the JobQueue to also record samples in the
in-flight database table. Add necessary logic to cope with another
instance currently processing that sample. (Resubmission of those samples
is currently tied to another analysis finishing which isn't optimal. We
need a separate thread here to regularly re-try acquiring those
samples.)

Update the db schema to 4 so we get fresh and empty tables.

Clear the current instance's in-flight samples from the database on
shutdown as well.

Do not save the analsysis info to the database upon registration of the
sample and do not update the sample info upon submission to Cuckoo. The
aim here seems to have been to be able to re-acquire Cuckoo jobs after a
Peekaboo restart which AFAIK never worked and which we have abandoned as
desired functionality for now in favour of getting concurrency right
first.