Remove sample object in-flight db connection wrappers
Now that the main queue has access to the database as well, there is no
need any longer to wrap this in the sample object. As a first step we
set and clear the in-flight markers of samples directly from the
queueing module and remove the respective wrappers from the sample
class.