Move result saving from sample to worker
Move the logic to save the analysis result from the sample to the worker
to make more transparent where database accesses are done. This also
highlights that while the actual checking whether a sample is known is
currently an optional rule the saving of analysis results is
non-optional. Add a FIXME note regarding that fact for now.