Remove DB schema version redundancy
Define schema version as global constant in DB module. Use it in table
names. Eliminate table names in foreign key definitons by using the
class names directly.

We use a global constant because a static class variable of
PeekabooDatabase is not visible when defining the database model classes
and we'd have to move the PeekabooDatabase class definition before them.
Making it a static class variable of PeekabooMetadata causes a type
error:

TypeError: %d format: a number is required, not InstrumentedAttribute

Does not seem worth the hassle just to avoid a global constant.