Rework rule result reason handling in Sample
Sample's reason property has only one user: analysis_save() of
PeekabooDatabase. We only save the analysis result after running the
ruleset and at the end of that RulesetEngine.report() which runs
Sample.report() which runs Sample.determine_result() which sets the
reason based on what the ruleset has done. So having a logic in the
reason property that goes looking for a reason in the database is
redundant.

We turn the reason property into a getter Sample.get_reason() to catch
other, undetected users and get in line with Sample.get_result(). The
reason returned by the getter will now be None for a freshly created
Sample because that reflects that no analysis has been done on it yet.

Removing the database lookup from the reason property in turn makes
PeekabooDatabase.fetch_rule_result() redundant so we remove that as
well.