Always save analysis job result
Previously we only saved the analysis result if we did not know the
sample already. This made the separate jobs table basically redundant
because there would only ever be one entry in analysis_jobs for the
first analysis done on a sample.

With this change we now save every analysis result even if we reused the
old, already known sample info. That way we log information about how
often we got to see the same sample and (based on the analysis
timestamp) in what timeframe.

The database module makes sure to re-use the already present sample info
when saving the new analysis result.

It should be noted that we still don't do anything useful with this
information.