Update developers list in documentation
Update the list of (active) developers in the documentation.

Intentionally leave all the authors references in the code alone because
these have not changed. It should be strongly considered removing them
or at least not polluting the code with even more since version control
is already tracking this data and much more granularly.

Closes #64.