Refactor report evaluation for submission error handling
If Cuckoo analysis fails we used to report the sample as bad, i.e. as
containing a virus which is not exactly true and might arguably be seen
as a false positive. Instead we should report the sample as failed
and let whoever called us deal with it however they want. This changes
sample reporting from a two-state (good, bad) logic to three-state
(good, bad, failed).

If Cuckoo itself reports to us that it was not able to complete the
analysis we can easily accomplish this by returning Result.failed
from rule cuckoo_analysis_failed.

For this we turn the hitherto unused Result.checked result into
Result.failed and reorder slightly. The numerical value here is
important to achieve correct ordering of result "worseness":

1. If any rule resports "bad" the sample has been positively identified
   as containing something malevolent and this should overrule any other
   finding.
2. If any rule reports "failed" then we don't know what it might have
   found and it could be "bad". So "failed" should be the next worst
   finding after "bad".
3. If any rule reports "good" then it positively identified the content
   as non-malevolent.
4. "unknown" should mean that we tried to identify the content but
   couldn't reach any conclusion either way.
5. "unchecked" should mean that we haven't even tried finding out
   anything about the sample (yet).

For cases where actual submission to Cuckoo failed we need to catch our
own CuckooAnalysisFailedException raised by Cuckoo.submit(). Since this
was hidden in the report acquisition code way inside Sample, we need to
move submission and report handling to the ruleset. Here we provide a
common base class CuckooRule implementing the evaluate() method in a way
that checks for the Cuckoo report to be present and submitting the
sample to Cuckoo if not. If the report is present (i.e. has been
reported back by Cuckoo later on) the actual rules' evaluate_report()
method is called by CuckooRule's evalute(), nicely modelling that those
rules actually only evaluate the Cuckoo report and not the sample
itself.

This by the way gets rid of the cuckoo_report attribute of sample,
replacing it with private member __cuckoo_report and a setter
Sample.register_cuckoo_report() called by Cuckoo.resubmit_with_report().