Move client interaction to frontend for overall verdict
Clients like AMaViS are interested in an overall verdict of all the
files presented to us in order to decide whether to forward the whole
email or quarantine it. We used to report on every single sample and
matching in AMaViS would consider the whole email as "bad" if one sample
was reported as "bad". Similarly, the whole email would be regarded as
"good" if no sample was reported "bad" and at least one reported any of
"unknown", "checked", "ignored", "good".

With the introduction of a third return value of "failed", this logic
breaks because AMaViS has no explicit matching for it. Instead, analysis
is considered to have failed if neither "good" nor "bad" was reported.
We do that but for each sample individually. So if all but one analyses
failed and the last returned "good" the whole email would be considered
"good" and forwarded even though we didn't and couldn't say anything
about all the other samples.

With this change we report an overall analysis result as well. The
format is chosen so that previous matching expressions for clients
match only this overall result and will thus correctly identify the
third case of at least one analsysis having failed.

We take the opportunity to rework client connection handling so that it
moves completely into the fontend. This eliminiates concurrent writing
of messages to the connection socket by the samples, potentially
garbling each other. We now record all the messages we want to report
to the client inside the sample (as we did anyway for dumping the
processing report) and send them to the socket from the connection
handler at the very end.

This also eliminiates the need for a connection map. Its final user, the
Cuckoo classes, can easily be converted to track the samples it has
submitted to Cuckoo for analysis by itself bu introducing a running_jobs
dictionary.