Move dump_processing_info to sample
dump_processing_info isn't really related to the ruleset engine but it
dumps lots of internal information of Sample. So make it a method of
Sample and call it from the worker as a similar postprocessing as saving
the result to the database.

Also dump the processing info if analysis failed for later analysis.