Rework stale pid file and socket handling
The server would try to start three times. This could have been to wait
for lingering connections to free up (EADDRINUSE) but was a.) missing an
appropriate sleep interval between tries and can b.) only happen with
TCP sockets, not UNIX domain. So it should be safe to remove it.

What it also did (implicitly) was clean up a socket left by a crashed
instance that *would* actually cause an EADDRIUSE because the socket
directory entry would already exist when the server started up.
Unfortunately, this would just as well remove the socket of a running
instance, allowing two instances to run in parallel and potentially get
in the way of each other.

Rework that logic to check for an existing PID file and a running
process with the PID mentioned in it (using os.kill(pid, 0)). Only start
up if either no PID file exists or no running process with that PID.
Clean up a stale PID file to get freash owner and permissions.

Similarly, explicitly check for an existing socket on startup and
whether someone is answering it. Only start up if no socket exists or
it is stale. Clean up a stale socket to get freash owner and permissions.

On shutdown, catch and warn about errors removing the PID file or socket
because they shouldn't go missing or change permissions while we're
running but shouldn't confuse us into not shutting down at all either.

Move the housekeeping logic into a new class
PeekabooDaemonInfrastructure for tidyness and automatic cleanup.