Consistently catch errors when talking to client
Add a method talk_back() to the request handler that catches errors when
talking to the client. This handles cases where the client would
arbitrarily close the connection. Showcased by our new stale socket
startup test where we open and close the socket right away.

Changes behaviour of regular update messages in wait(): Previously we
would try to keep running when there was an error other than EPIPE when
talking to the client. This doesn't seem to make much sense.