Localise client communication
Since we're communicating with a client that potentially wants to
present our findings to an administrator or even end user to inform them
about what's going on with their email or data, we had German messages
going to the client. Obviously, this is very inconvenient and inflexible
for non-German speakers. Also, it doesn't look very good in the code
base.

Replace static German strings in client communication with English
message ids for internationalisation by gettext. Add an appropriate
German message file. Use the pure-python standard-library gettext module
of python to employ message catalog files. Add the catalog files to
MANIFEST.in for automatic installation into the package by setup.py. Add
Babel as a development dependency to extract messages and update and
compile the catalogs. Add documentation for developers and translators,
how to work with these translations.