Report failure on rule failure
If a rule failed by exception, e.g. because communication with the
Cuckoo REST API failed due to networking issues, we would log a message
and return result "unknown" for that rule and continue processing. This
led to false positives where samples would still be classified even
though we had an obvious processing failure and shouldn't classify them
at all.

Change behaviour to report result "failed" and abort rule processing.
This will report the analysis as failed to the client and let them
decide wether to resubmit, quarantine or pass on.