Avoid encoding errors using encoding-aware open
Writing the report file in dump_processing_info() would throw exceptions
such as:

UnicodeEncodeError: 'ascii' codec can't encode character u'\xfc' in
position 37: ordinal not in range(128)

if anything in the report was not plain ascii, such as an input file
name containing umlauts. Solve this by importing the open function from
the futures builtins module, i.e. use python3 semantics where files
implicitly use the system locale for input and output encoding. write()
then accepts whta is unicode() in python2 and plain strings in python3
and automatically encodes them for the output file's encoding.

As a side-effect this throws up a problem where the json module insists
of returning a bytearray (or rather str() in python2) which write will
not accept with:

TypeError: write() argument 1 must be unicode, not str

This can be side-stepped nicely by opening the JSON dump file with flag
'b'. But this then fails on python3 because that one expectes a
byte-like object for binary output which we produce by explicitly
encoding to 'ascii'. This works because json by default ensures its
output is ascii clean. But for paranoia's sake we add the ensure_ascii
option anyway. This workaround (including the 'b' on open) can go away
when we stop supporting python2.