Leave DB engine disposal to garbage collector
Explicitly disposing of the SQLAlchemy engine in __del__ of
PeekabooDatabase causes an exception with python3:

Exception ignored in: <bound method PeekabooDatabase.__del__ of
<peekaboo.db.PeekabooDatabase object at 0x7f7c335ed160>>
Traceback (most recent call last):
  File "peekaboo/db.py", line 465, in __del__
  File "lib/python3.6/site-packages/sqlalchemy/engine/base.py", line 2014, in dispose
  File "lib/python3.6/site-packages/sqlalchemy/pool/impl.py", line 251, in recreate
  File "<string>", line 2, in __init__
  File "lib/python3.6/site-packages/sqlalchemy/util/deprecations.py", line 130, in warned
  File "lib/python3.6/site-packages/sqlalchemy/pool/base.py", line 221, in __init__
  File "lib/python3.6/site-packages/sqlalchemy/event/base.py", line 150, in _update
  File "lib/python3.6/site-packages/sqlalchemy/event/attr.py", line 367, in _update
  File "lib/python3.6/site-packages/sqlalchemy/event/registry.py", line 115, in _stored_in_collection_multi
KeyError: (<weakref at 0x7f7c37bd2e58; to 'function' at 0x7f7c335b3378 (go)>,)

Best guess: SQLAlchemy is working with weakrefs
(https://docs.python.org/3/library/weakref.html) to get around problems
which might prevent proper garbage collection (like cyclic dependencies
between objects) and in our case this backfires particularly because
we're doing cleanup when garbage collection is already underway.

Doing complex cleanup work in __del__ is not generally recommended, most
notably because it is highly uncertain when garbage collection will
actually happen.
(https://stackoverflow.com/questions/1481488/what-is-the-del-method-how-to-call-it)

Since we've been doing the disposal only on garbage collection anyway
and have wrapped the engine inside the PeekabooDatabase object so that
new instances also get a new engine, we can just as well also rely on
garbage collection to also dispose of the engine of a disposed
PeekabooDatabase.

If there are problems down the road, an explicit close method would be
the way to go if reliable cleanup is required at a specific point in the
code.