Remove explicit references to python 2
We no longer care which python version we run under. So remove explicit
references to python2 and just let the tools and user choose what they
like.

The one exception is the interpreter override for embedded mode. Here we
change the default to /usr/bin/python2 to force that version for
execution of Cuckoo since it doesn't run with python3 yet.