Let requests and urllib3 do our retries for us
Use a session configured with various retry parameters and a special
whitelisting retry class do our retries for us. This offloads all kinds
of error handling we've not even thought of yet to urllib3 while still
staying highly customizable. This also gives us a nice backoff algorithm
when retrying which we only need to parametrize.

While at it, unroll the __get() method into the one user of POST
(submit()) to make it more readable.