Distribute sample configuration files
Add ruleset.conf.sample to the distribution and configure setuptools to
install it together with the README, peekaboo.conf.sample and our
systemd unit into an fhs-like documentation directory.

Add a number of comments to setup.py so we remember why this seems to be
the least disruptive way out of the limited options presented by the
python/setuptools/distutils/pip mechanics of providing these files at a
location visible to users and still somewhat manageable by alternative
packagers (e.g. Linux distributions).

In short again:
- package_data would hide the files away below the site-packages
  directory making them invisible to the user and hard(er) to find
  programmatically
- absolute paths really don't work well with data_files
- relative paths work somewhat better but still not fully consistently
  between setup.py and pip
- overriding the install command solves the installation but not the
  packaging problem