Add generic rule on cuckoo report to test for matched cuckoo signatures
Add IdentifierMissingException in expressions to notify about identifiers
not included in the given context e.g. cuckooreport. This way the report
can be requested on demand, sumbitted, added to the context and evaled
again on demand.
Add signature_descriptions to cuckoo module in toolbox to get direct
access to signature descriptions from generic rules.
Add test to check generic rule on cuckoo report.