oletools: Add bug workaround for macro extraction
When called on an empty or text document oletools will falsely report that it
contains macros and returns a one item list of macros which contains only the
filename again.

Add a workaround to the oletools analyser to detect this behaviour, warn about
it and override the result.

Until a fixed version of oletools is available.

Extend the test suite to catch these cases.