expressions: Make evaluation objects stateless
Remove the feedback mechanism which has only one user and adds lots of
internal state to the evaluation objects. Similarly, drop the now
unnecessary subeval() helper which also had only one user, giving a much
clearer picture of what the operator implementors are doing.

Apart from simplification this is in preparation for switching to a
single, global set of pre-parsed expression-based rules which we only
ever evaluate against changing contexts but which are otherwise
read-only and stateless. This in turn apart from being more elegant and
efficient should pave the way to handle the non-thread-safety of
pyparsing.