ruleset: Simplify expressions rule identifier resolution
Make identifier resolution in expressions rule an endless loop which we
immediately break out of if rule.eval() did not throw an exception. Add
some clarifying comments to that effect.

Additional statements in the loop need to either return or break out of
it to prevent an endless loop. Before they could have set result to
something but that would have been even more confusing.