oletools: Avoid repeated report construction
Fold repeated construction of report objects together for simplification
and reduced potential for regressions. In particular, uset
Rule.get_oletools_report() in OleRule as well.

Eliminates (maybe left-over?) Exception catching and re-raise of unclear
benefit.

With recent changes to ole.py, the OletoolsReport can no longer be None
but will instead be an empty report on error which certainly simplifies
the rules.