ruleset: Pre-init for efficiency and thread-safety
Previously, we would reinit the whole ruleset at every evaluation. Apart
from being inefficient this also exposed thread-safety issues in
pyparsing.

Switch to initialising one global RulesetEngine with all the rules
constructed, validated and parsed. Since this happens in the main thread
at startup, it should avoid all thread-safety issues. This greatly
simplifies the ruleset engine with more potential through refactoring.

This also means that the ruleset engine and individual rules are no
longer allowed to change their internal state. They did not do so
before, so this should not be a problem. A cautioning comment in
Rule.evaluate() hopefully reaches rule implementors.

With this change, config and db_con parameters for rules no longer need
to be optional. Adjust the testsuite to pass explicit Nones for now
which should make us catch changes later on where a rule starts to
actually use that parameter.

Closes #110.