Kill the debugger
Having a module whose stated purpose is to execute arbitrary code in the
context of a security solution active in released versions and only a
config switch away is a recipe for disaster. There are any number of
other ways to debug problems while coding, so many in fact that I have
not used it ever. Kill it for good.