expressions: Improve missing identifier handling
Currently we rely on the internal workings of KeyError and order of its
args array to get at the name of the identifier we need to provide.
Turns out, exceptions can carry custom attributes. Use that to store the
exact identifier name and provide a nice error message in the process.
Rework value retrieval slightly to use more speaking names.

Add a testsuite test to catch regressions.