expressions: Catch incomplete context through KeyError
EvalIdentifier would report a missing 'variables' key of the context as
a missing identifier. Since this is not really the case, modify it to
throw an actual KeyError so we are made aware of any calls where it's
missing. Add a test to the testsuite to catch regressions on this.

Also, since we do know exactly what identifier we were asking for there
is no need to get it out of the KeyError exception.