expressions: Don't do work in exception handler
Identifier retrieval seems to tend to lock up in Oletools under high
load and only under python2. While this is not the solution, it does not
seem a good idea to do this much work in an exception handler.

Move identifier out of the exception handler.

(While the diff looks daunting it's just an unshift of the whole block.)