server: Correctly handle submit error
Commit 1d1ba965a3 factored sample submission out of request parsing to
allow null requests and eventually multiple different request types.
This refactoring did not take into account that the method needs to
return None or at least an empty list if an error occured. Instead, None
returned on error by submit_sample() was unconditionally added to the
submitted sample list, making it non-empty and thus non-false. This lead
to wait() being called on that list of Nones which would never return
and leave the client connection starving.

Restore consistent behaviour by aborting if a sample fails to be
submitted.

Closes #122.