db: Add column indices for efficiency
We've seen high CPU load on the database cause by Peekaboo. This is
because we've totally forgotten to add any indexes other than what the
DBMS automatically does for primary keys. This makes our database
operations very inefficent which comes to bear once the sample info
table contains a couple of hundrets of thousands of entries and is
aggravated by a large number of instances in cluster mode.

Add database indices to speed up processing and lower database CPU
load. Raise schema version to 7. Add comments in the code explaining
what the considerations are so a coder can reevalute on future schema or
code changes.

Closes #124.