rules: Silence expressions debug logging
Currently the expressions module will spam the logs with verbose
debugging if the global loglevel is set to DEBUG. Instead of removing
debug logging from the module, introduce a rule-specific log_level
configuration option with a default of WARNING. That way we silence log
spamming by the module but can still enable additional logging if
necessary. For simplicity, we just configure the module's logger
directly from the rule instead of propagating object-specific loggers
around.