rules: Catch invalid expression rules
Logging a warning and returning failed during evaluation if some
expression is iffy is a good fallback but not very helpful to the admin.

Extend eval objects in expression parser so they can be interviewed
whether they are expressions and use that to error out on startup if the
expression does not fulfil our requirements of an expression rule.

Somewhat overly specific in implementations but it's our parser so no
reason not to make it do what we want. Can and should be generalised
later on as necessary