systemd: Start after mysql or postgres
Similarly to the cuckoo api service we want to be started after our
database is up. Since we support mysql and postgres (and sqlite which
does not have a daemon), we add ordering for both of them. Since we
don't know which the user chooses, we're not adding a hard dependency on
either one. Instead the admin can enable the respective service or add a
Requires= directive in a local systemd drop-in config.

Possible fallout from this is that a user has mysql *and* postgres on a
peekaboo machine and for some reason wants to start the one not used by
peekaboo after peekaboo. Our ordering would break this. But this should
also be overrideable by a drop-in config.