cuckoo: Make submission with original filename optional
Submission to Cuckoo with the original filename is problematic under
some circumstances. Some are purely technical and arguably bugs that
should be fixed (such as encoding problems), others are operational
(such as intentionally not leaking filenames into the Cuckoo
environment). Either way we should give users the option to choose to
work around technical problems or implement their policy own on original
filenames.

Add a config option to switch submission to Cuckoo with the original
filename on and off.