sample: Always obfuscate original filename
Previsously we would only obfuscate the original filename if we could
determine an extension. This was an overly enthusiastic optimisation
that undermined original filename obfuscation towards Cuckoo. Restore
the previous behaviour of always using the sha256sum as basename and
only leaking the file extension because that's required by Cuckoo to
determine the analsysis package (or is it?).